requirements.txt
openai==0.28.0
gradio
langchain
python-dotenv
black
isort

faiss-cpu
tiktoken
spacy